WIP: implement `sret_union` ABI for pointer-ful types
This is a combination the existing:
 - `sret`  ABI (which can stack-allocate a _single_ pointerful type)
 - `union` ABI (which can stack-allocate many _pointer-free_ types)

This provides some nice speed-ups for temporary "wrappers":
```julia
const v = Any[]
@noinline maybe_wrapped(i) = (i % 32 != 0) ? Some(v) : nothing
function foo()
    count = 0
    for i = 1:1_000_000
        count += (maybe_wrapped(i) !== nothing) ? 1 : 0
    end
    return count
end
```

On this PR this gives:
```julia
julia> @btime foo()
  1.675 ms (0 allocations: 0 bytes)
968750
```

compared to current master:
```julia
julia> @btime foo()
  6.877 ms (968750 allocations: 14.78 MiB)
968750
```

The most outstanding TODO here is what to do about PHI nodes. Right now,
if the incoming `Union{...}` type has a pointer-containing type then the
object is forced to be boxed, even if the object at run-time is actually
pointer-free.

But that's just a band-aid - it means we introduce new boxes where we
didn't have them before, which is a regression that almost certainly to
be fixed before landing this.

Co-authored-by: Gabriel Baraldi <baraldigabriel@gmail.com>