Add `_call_assert_precompiled` intrinsic
The purpose of this is to experiment with an "escape hatch" for `juliac`
see that will allow you to annotate your code at specific dynamic
dispatch sites, to say that you've verified that the types being passed
in here are covered by `precompile()` statements elsewhere.

These call-sites are uninferable, so it highly recommended to use this
with a type-assertion. For example:
```
while !eof(f)
    n += @assert_precompiled readline(f)::Int
end
```