Do not give `StringVector`s to users
A `StringVector` v is a `Vector{UInt8}` which wraps memory allocated as a string.
This allows `String(v)` to be zero-copy which is good for performance.
However, if users are handed a `StringVector` directly, they can use that to
unknowingly mutate strings, which is illegal.
For example, suppose a user calls a function `f` which returns a `StringVector`
`v`. They can then do the following to mutate a string `s`:

```julia
v = fill!(Base.StringVector(3), 0x00);
v2 = reshape(v, 3, 1)
s = String(v) # all zeros
v2[1] = 0x01
s
```

The solution in this PR is to not ever let `StringVector`s escape out to the
user, by instead returning normal `Vector{UInt8}` from the functions that would
otherwise do so.