Try giving some more informative errors for various malformed input
I was playing with generating some code for OpaqueClosures. It's
pretty easy to generate IR that will pass the verifier, but cause
assertion errors in codegen. This tries to make the experience
slightly nicer by turning some of them into proper error messages
(thus letting the runtime discover, so that e.g. the code can be
inspected at the REPL) rather than assertions.