Specialize `copyto!` for triangular matrices (#52730)
This provides a performance boost in copying a triangular matrix to a
`StridedMatrix`, which is a common operation (e.g. in broadcasting or in
`Matrix(::UpperTriangular)`). The main improvement is improved cache
locality for strided triangular matrices by fusing the loops.

On master
```julia
julia> U = UpperTriangular(rand(4000,4000));

julia> @btime Matrix($U);
  64.649 ms (3 allocations: 122.07 MiB)
```
This PR
```julia
julia> @btime Matrix($U);
  48.332 ms (3 allocations: 122.07 MiB)
```