Use Base.require() to load Distributed internally
Base.require_stdlib() will exclusively load a stdlib from whatever was shipped
with Julia. The problem is that when developing Distributed.jl this will cause
the workers to always load the builtin Distributed module instead of the
development version, which can break everything because now the master and
worker are running different (potentially incompatible) versions of
Distributed.

This commit changes Base and LinearAlgebra to use `Base.require()` instead,
which will respect `Base.LOAD_PATH`. I argue that this is safe because unlike
the other stdlibs like REPL or Pkg, Distributed is only loaded if explicitly
requested by the user with `-p` or through `addprocs()` or something, so it
shouldn't be possible to get into quite the same
tools-are-broken-because-I-broke-my-tools situation that motivated using
`Base.require_stdlib()` in the first place.

An alternative design would be to:
1. Move the if-block loading Distributed in `exec_options()` below the for-loop
   where it will execute `-e` options.
2. Require any implementation of `Distributed.launch(::ClusterManager)` to pass
   `-e 'using Distributed'` in their command to ensure that Distributed is
   loaded in a way respecting `Base.LOAD_PATH`.

This would be more consistent with how the other stdlibs must be developed, but
it requires implementers (i.e. Distributed and ClusterManagers.jl) to opt-in to
allowing development versions of Distributed, which feels very annoying and easy
to miss so I decided not to implement that.