Tweak malloc_stack on OpenBSD to match other implementations; explain lack of guard page (#54564)
The other malloc_stack implementations were switched to using
jl_atomic_fetch_add_relaxed on 2024-02-23. I suspect that this simply
overlapped in time with the creation of the OpenBSD patches, which were
submitted in March, but presumably work started before that :-).

CC @semarie please check that I am not messing up something for you.

Also, while at it, I noticed that the OpenBSD `malloc_stack` is missing
the `mprotect` for setting up a guard page, which concerns me a bit. Is
there a deeper reason for this omission?