Don't expose guard pages to malloc_stack API consumers
Whether or not a guard page is in effect is an implementation
detail and consumers of the `malloc_stack` API should not have to
worry about that. In particular, if a stack of a certain size is
requested, a stack of that size should be delivered, and not be
reduced on some systems because we park a guard page in that
range.

This also helps consumers of the gcext API implementing stack
scanning (i.e., GAP.jl), as it does not have to worry about
running into those guard pages anymore.