codegen: start to remove the ability to call back into inference
This is preparation for making codegen purely a transform of CodeInfo ->
Module* without involving any of the caches (at the C++ level), so that
this can instead be driven by caches managed entirely at the Julia
inference level.

Removes the test for code_llvm lookup behavior, since codegen no longer
does lookup, so that is no longer relevant to test.