Add boundscheck in bindingkey_eq to avoid OOB access due to data race (#54671)
The race here is that svec might be replaced and a new binding
introduced into the keyset while we hold a reference to the old svec,
which led to a OOB access on the svec with the index a binding
introduced at the same time. This now introduces a bounds check which
will force taking the lock if we fail the lookup i.e we had a data race.

Fixes https://github.com/JuliaLang/julia/issues/54285

---------

Co-authored-by: Jameson Nash <vtjnash@gmail.com>
(cherry picked from commit 20f03ddcefffaf5e42d9060144a6cc3aa1313dd4)