Force generated functions to have min_world == 1 if they are unbounded and have no edges
Our pkgimage generation code uses `min_world == 1` as a shorthand for
"no backedges target this" to skip revalidation on load, which I ran into
when I accidentally broke that assumption in #54654. However, I then realized
that is was possible for this property to be false, even today in the presence
of generated functions. The failure case isn't too terrible (it just fails to use
the precompile'd version), but it's still a very subtle silent bug. For now,
forbid this configuration in generated functions so that the staticdata assumption
is always valid and add a test case that would have failed with #54654 to make
sure we notice if this regresses.