Force @generated to have min_world == 1 if they are unbounded and have no edges (#54699)
Code instances are supposed to have `min_world == 1` if and only if
no backedges target the ocde instance. The package image code in particular
uses this to to skip revalidation on load. I ran into this issue
into when I accidentally broke that assumption in #54654. However, I
then realized that is already possible for this property to be false, even
today in the presence of generated functions. The failure case isn't too
terrible (it just fails to use the precompile'd version), but it's still
a very subtle silent bug. For now, forbid this configuration in
generated functions so that the staticdata assumption is always valid
and add a test case that would have failed with #54654 to make sure we
notice if this regresses.