lowering: Remove `outerref` intermediate form
The `outerref` form was added in c3eedce56e6f7759873d7c33ef67420c7bc4fe77, but
the renaming pass for them was removed two years later in c44644442949238da70670b547312ca1ae9a9c7d,
without removing the form itself. As far as I can tell, today, `outerref`
is essentially equivalent to `(globalref (thismodule) name)` (which was
the original form, before outteref was introduced) because it just
expands to a toplevel symbol. I don't think there remains any reason
to keep this form after the renaming pass was removed and moreover it
is confusing as one could reasonaly infer (as I did an incorrectly
wrote in a comment) that `outerref` provides outer access, which
it does not. This PR removes the form entirely and replaces them
with an appropriate globalref.