Remove unused `JL_HAVE_SIGALTSTACK` code (#54568)
Nothing is ever activating this, and it probably hasn't been tested in a
while.

As far as I can tell, this code path was added by @vtjnash in
082d7d2544e8e080d5f260c94d093480ee72e360

I have no particular beef with this code other than "it complicates the
thread / task logic, so dealing with this requires understanding one
more moving part, removing it makes it simpler". If it is decided that
this code should rather be kept, "just in case" (which, though), that's
equally fine by me. Thought perhaps a comment could be added explaining
the motivation for keeping it even thought it is currently always
disabled?