[TOML] remove Dates hack, replace with explicit usage (#54755)
This hack will cease functioning soon (#54739), so it must be removed so
that packages can stop relying on it.

As an aid, now provide some basic conversion and printing support for
the Base.TOML variants of this as well such that users can round-trip
the contents (if they are well-formed as a date/time).

(cherry picked from commit 2ce12e965c4da2c042343cf6258aab9d7c7f4065)