Record macro expansions in expression provenance
Here we track expression provenance as a tree. Some operations like
macro expansion and quote expansion have multiple sources of provenance
for a single expression - here this is represented as a tuple of nodes
in the `source` attribute.

This allows us to extract the macro expansion stack for any expression
by walking the provenance tree; we use this to fill in the proper debug
info when converting to CodeInfo.