Follow up #54772 - don't accidentally put `Module` into method name slot
The `:outerref` removal (#54772) ended up accidentally putting a `Module`
argument into 3-argument `:method` due to a bad refactor. We didn't catch
this in the tests, because the name slot of 3-argument `:method` is
 unused (except for external method tables), but ordinarily contains
the same data as 1-argument `:method`.
That said, some packages in the Revise universe look at this
(arguably incorrectly, since they should be looking at the
signature instead), so it should be correct until we fix Revise,
at which point we may just want to always pass `false` here.