block loading extensions when the package is not yet requested
When loading from require_stdlib, the code is supposed to be inserted
directly into the unrequested state only. This means that extensions
should not automatically run, as these modules do not "interact" unless
required to by a package that depended upon both, or if the user later
explicitly imports them.

Fixes #54884