cglobal: Fall back to runtime intrinsic
Our codegen for `cglobal` was sharing the `static_eval` code for symbols
with ccall. However, we do have full runtime emulation for this intrinsic,
so mandating that the symbol can be statically evaluated is not required
and causes semantic differences between the interpreter and codegen, which
is undesirable. Just fall back to the runtime intrinsic instead.