Fix potential underrun with annotation merging
As raised in #54860, when writing to an AnnotatedIOBuffer, should the
new content have more annotations than the AnnotatedIOBuffer, we may
attempt to index non-existent annotations.

This bug occurs in the process of looking for runs of matched
annotations. Since it is impossible for a run to be longer than the
number of existing annotations, we can add this as a sanity check and
not bother trying to check for run lengths where this is not the case.

Reported-by: caleb-allen <caleb.e.allen@gmail.com>