Test system for testing IR
New system for testing generated IR based on pretty printing the IR and
comparing to a reference string. String comparison is an excellent
way to do IR comparisons because
* It's very simple to implement
* Writing test cases is simple
* It's easy to pinpoint errors in tests
* Irrelevant detail can easily be omitted in the pretty printing

The main downside is that, in principle, the pretty printing may have
ambiguities, and changing the pretty printing causes high churn in test
cases. However these seem like reasonable tradeoffs, given the benefits.