add mechanism for configuring system image builds (#54387)
This adds the option to pass a filename of configuration settings when
building the Core/compiler system image (from
`base/compiler/compiler.jl`). This makes it easier to build different
flavors of images, for example it can replace the hack that
PackageCompiler uses to edit the list of included stdlibs, and makes it
easy to change knobs you might want like max_methods.