Fix accidental early evaluation of imported `using` binding
In `using A.B`, we need to evaluate `A.B` to add the module to the
using list. However, in `using A: B`, we do not care about the value
of `A.B`, we only operate at the binding level. These two operations
share a code path and the evaluation of `A.B` happens early and is
unused on the `using A: B` path. I believe this was an unintentional
oversight when the latter syntax was added. Fixes #54954.