Color todo admonitions in the REPL
This patch adds magenta coloring for `!!! todo` admonitions in addition
to the existing styling of `danger`, `warning`, `info`, `note`, `tip`,
and `compat` admonitions. This is useful if you want to leave some more
colorful todo notes in docstrings, for example.

Accompanying PR for Documenter to render these in HTML and PDF docs:
https://github.com/JuliaDocs/Documenter.jl/pull/2526.