module: Prepare `usings` array for world age partition
This is a relatively independent part of the bindings partition branch,
extending the module usings list to gain `min_world` and `max_world`
`size_t`s. These are always `0` and `(size_t)-1` respectively in this
PR, which handles the GC and serialization implications of this change.
The semantic part will come later.