staticdata: Unique Bindings by mod/name (#54993)
Currently we error when attempting to serialize Bindings that do not
beloing to the incremental module (GlobalRefs have special logic to
avoid looking at the binding field). With #54654, Bindings will show up
in more places, so let's just unique them properly by their module/name
identity. Of course, we then have two objects so serialized (both
GlobalRef and Binding), which suggests that we should perhaps finish the
project of unifying them. This is not currently possible, because the
existence of a binding object in the binding table has semantic content,
but this will change with #54654, so we can do such a change thereafter.