Fix a regression in the test for #13432 (#55004)
The test for #13432 is supposed to test a particular codegen path
involving Bottom. It turns out that this code path is regressed on
master, but hidden by the fact that in modern julia, the optimizer can
fold this code path early. Fix the bug and add a variant of the test
that shows the issue on julia master. Note that this is both an
assertion failure and incorrect codegen. This PR addresses both.