add missing setting of inferred field when setting inference result
This previously could confuse inference, which expects that the field is
set to indicate that the rettype has been computed, and cannot tolerate
putting objects in the cache for which that is not true.

Also cleanup `sv.unreachable` after IR modification so that it remains
(mostly) consistent, even though unused (except for the unreachable
nodes themselves), as this was confusing me in debugging.